Solo falta el deploy

+++ src/main/java/com/bd2/proyecto1/deleteOfertaServlet.java
@@ -33,14 +33,27 @@ public class deleteOfertaServlet extends HttpServlet {
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
-       
+
         OracleDBConnection db = dbSingleton.getDBConnection();
         ofertaRepository ofeDB = db.getOfeRep();
+        int isPossible = ofeDB.isMine(db.getUserLogged().getId(), Integer.parseInt(request.getParameter("oferta-borrar-form-id")));
+        if (isPossible == -1) {
+            request.getRequestDispatcher("noEsTuyo.jsp").forward(request, response);
 
-        ofeDB.delete(Integer.parseInt(request.getParameter("oferta-borrar-form-id")));
+            return;
+        }
+        db = dbSingleton.getDBConnection();
+        ofeDB = db.getOfeRep();
+
+        int done = ofeDB.delete(Integer.parseInt(request.getParameter("oferta-borrar-form-id")));
         db.closeConection();
 
-        request.getRequestDispatcher("comprarOVender.jsp").forward(request, response);
+        if (done == -1) {
+            request.getRequestDispatcher("falloPagina.jsp").forward(request, response);
+            return;
+        }
+
+        request.getRequestDispatcher("exitoPagina.jsp").forward(request, response);
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
